refactor: Extract runner.go into separate files by responsibility
Split the monolithic runner.go (1400+ lines) into focused modules:
- directive.go: Directive parsing and task metadata extraction
- hooks.go: Git hooks and shell command execution
- interaction.go: User interaction dispatch and waiting
- prompt.go: System prompt construction
- tool_description.go: Tool description formatting
Also rename runner_test.go to directive_test.go to match the new structure.
